lkl: Remove gettid from host_ops
ost_ops.gettid is not called anywhere except in a boot test. The boot
test just checks to see if gettid returns a non-zero tid, which makes no
sense.

I suspect host_ops.gettid was added for some historical reason, but it
can now be safely removed. This commit removes it as well as the silly
test.

Signed-off-by: Ruihan Li <lrh2000@pku.edu.cn>