lkl: Run all IRQs in lkl_trigger_irq
The main body of lkl_trigger_irq seems to run only the specified IRQ,
but all pending IRQs are run in lkl_put_cpu anyway. If the current IRQ
is already pending, the same IRQ is run twice.

This commit changes the logic to mark the current IRQ as pending and
then try to run all pending IRQs, which should be clear logic.

Signed-off-by: Ruihan Li <lrh2000@pku.edu.cn>